+++ source/app/MLValue.cpp
@@ -108,10 +108,10 @@ Value::Value(int v) : mType(kFloatValue) { mFloatVal = v; }
 
 Value::Value(bool v) : mType(kFloatValue) { mFloatVal = v; }
 
-Value::Value(unsigned long v) : mType(kUnsignedLongValue) { mUnsignedLongVal = v; }
+Value::Value(unsigned long v) : mType(kUnsignedLongValue) { mUnsignedLongVal = static_cast<uint32_t>(v); }
 
 // truncate to unsigned long for now. 
-Value::Value(unsigned long long v) : mType(kUnsignedLongValue) { mUnsignedLongVal = static_cast<unsigned long>(v); }
+Value::Value(unsigned long long v) : mType(kUnsignedLongValue) { mUnsignedLongVal = static_cast<uint32_t>(v); }
 
 Value::Value(uint32_t v) : mType(kUnsignedLongValue) { mUnsignedLongVal = v; }
